Implementing a big bunch of sortedList.h fixes & improvements:
Including <stdexcept> for std::out_of_range.
Removing redundant typedef from the node_t struct and instead just
naming it Node.
Adding a constructor to Node that receives data and initializes next to
be nullptr.
Modifying insert to use Node's new constructor in place of the original
manual initialization.
On that subject: changing all instances of NULL to nullptr.
Removing the freeList instance method in favor of a static one. (Do
consider making it a static method of Nod, though, I suppose. Same for
copyList.)
Adding a bunch of const qualifiers to methods that don't modify the
list, as well as to parameters that won't be modified.
Implementing a basic version of the filter & apply methods, and
replacing their function parameter with a generic functor.
Making the remove method work with a reference of the iterator rather
than a copy of it.
Letting the const_iterator use the defaut destructor.
Fixing several instances of using -> instead of . to access a
SortedList&'s list.
Adding an out_of_range exception throw if attempting to remove an item
pointed by an iterator that reached the list's end.
Prepending "typename" to several places (mainly method implementations)
that require it.
Allowing const_iterator's operator++ to take it all the way to nullptr
rather than the node before it, otherwise we're skipping the last
element when iterating.
Making const_iterator's operator* return a constant T reference rather
than a constant Node reference.
Modifying findPreviousElementPosition to use >= rather than > to make it
work with the "remove" method (otherwise it'd remove the first element
*after* the one slated for removal).
Implementing a copyList method (but not using it yet).